fix(firebase and layout): adjust a number of hrefs, fix images components, re-istblish firebase
#14
