--- src/pages/authentication/index.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-function init() {
-    return <div>signin and signup pages</div>;
-}
-
-export default init;
