+++ src/pages/events/index.jsx
@@ -2,8 +2,15 @@ import React from "react";
 
 import EventList from "@/components/eventsPageComponents/EventList";
 
+import Layout from "@/layout/Layout";
+
 function index() {
-    return <EventList />;
+    return (
+        <Layout>
+            {" "}
+            <EventList />
+        </Layout>
+    );
 }
 
 export default index;
